Refactor: new files structure for better readability; add `index.ts` for easier imports
