+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/index.ts
@@ -1,0 +1,17 @@
+// UI Components
+export { default as FlipCard } from './ui/FlipCard'
+export { default as ErrorBoundary } from './ui/ErrorBoundary'
+
+// Wallet Components
+export { default as ConnectWallet } from './wallet/ConnectWallet'
+export { default as Account } from './wallet/Account'
+export { default as Transact } from './wallet/Transact'
+
+// Sunset Components
+export { default as HeroSection } from './sunset/HeroSection'
+export { default as MapFullScreen } from './sunset/MapFullScreen'
+export { default as PhotoUpload } from './sunset/PhotoUpload'
+export { default as Leaderboard } from './sunset/Leaderboard'
+
+// Layout Components
+export { default as Header } from './layout/Header'
