+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/sunset/HeroSection.tsx
