+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/sunset/PhotoUpload.tsx
