+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/ui/ErrorBoundary.tsx
