+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/ui/FlipCard.tsx
