+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/wallet/Account.tsx
@@ -1,7 +1,7 @@
 import { useWallet } from '@txnlab/use-wallet-react'
 import { useMemo } from 'react'
-import { ellipseAddress } from '../utils/ellipseAddress'
-import { getAlgodConfigFromViteEnvironment } from '../utils/network/getAlgoClientConfigs'
+import { ellipseAddress } from '../../utils/ellipseAddress'
+import { getAlgodConfigFromViteEnvironment } from '../../utils/network/getAlgoClientConfigs'
 
 const Account = () => {
   const { activeAddress } = useWallet()
