+++ QuickStartTemplate/projects/QuickStartTemplate-frontend/src/components/wallet/ConnectWallet.tsx
